# internal/api/docs/swagger.yaml
basePath: /
definitions:
  main.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  main.GroupListResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/model.Group'
        type: array
    type: object
  main.createGroupRequest:
    properties:
      name:
        type: string
    type: object
  main.createInstanceRequest:
    properties:
      cpu:
        minimum: 1
        type: integer
      gpu:
        type: integer
      name:
        type: string
      ram:
        minimum: 1
        type: integer
    required:
    - cpu
    - name
    - ram
    type: object
  model.Group:
    properties:
      instances:
        items:
          $ref: '#/definitions/model.Instance'
        type: array
      k3s_namespace:
        type: string
      name:
        type: string
      openstack_project_id:
        type: string
    type: object
  model.Instance:
    properties:
      cpu:
        type: integer
      external_id:
        description: k3s pod UID or openstack server ID
        type: string
      gpu:
        type: integer
      id:
        type: string
      name:
        type: string
      provider:
        description: '"k3s" or "openstack"'
        type: string
      ram:
        type: integer
      status:
        type: string
    type: object
info:
  contact: {}
  description: Demo API for multi-cloud Group and Instance orchestration.
  title: Orchestrator API
  version: "1.0"
paths:
  /groups:
    get:
      description: Returns all Groups.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GroupListResponse'
      summary: List groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a new Group (namespace in k3s + project in OpenStack).
        Rolls back on partial failure.
      parameters:
      - description: Group payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Group'
        "400":
          description: Invalid or missing name
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "409":
          description: Group already exists
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Cloud provider error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create a group
      tags:
      - groups
  /groups/{name}:
    get:
      description: Returns one Group by its name, including its instances.
      parameters:
      - description: Group name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Group'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get group by name
      tags:
      - groups
  /groups/{name}/instances:
    post:
      consumes:
      - application/json
      description: Creates a new Instance. GPU > 0 routes to k3s (pod), otherwise
        to OpenStack (server). Flavor is auto-matched.
      parameters:
      - description: Group name
        in: path
        name: name
        required: true
        type: string
      - description: Instance specs
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Instance'
        "400":
          description: Invalid or missing fields
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Unknown provider
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Cloud provider error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create an instance in a group
      tags:
      - instances
swagger: "2.0"